+++ webapp/src/pages/dashboard/order/[id]/index.tsx
@@ -119,7 +119,10 @@ export default function OrderObizPage() {
   const handlePDFActions = async (isShare: boolean) => {
     if (typeof order.ticket === "object" && order.ticket?.url) {
       try {
-        const response = await fetch(order.ticket.url);
+        const response = await fetch(
+          `/api/image?filename=${order.ticket.filename}`
+        );
+        console.log(`/api/image?filename=${order.ticket.filename}`);
         const blob = await response.blob();
         const filename = `bon-${order.offer.partner.name}-${order.number}.pdf`;
 
@@ -324,7 +327,7 @@ export default function OrderObizPage() {
         minH="full"
         direction={"column"}
         position="relative"
-        pt={16}
+        pt={8}
         pb={9}
         px={6}
         bg="bgGray"
@@ -338,6 +341,7 @@ export default function OrderObizPage() {
             rounded={"2xl"}
             gap={3}
             mx={4}
+            mt={-6}
           >
             <Flex alignItems={"center"} gap={2}>
               <PartnerImage
@@ -418,14 +422,14 @@ export default function OrderObizPage() {
               <Text>{formatDateToDDMMYYYY(order.createdAt)}</Text>
             </Flex>
             <Divider my={4} />
-            <Text
+            {/* <Text
               fontWeight={700}
               textDecor="underline"
               textDecorationThickness="2px"
               textUnderlineOffset={2}
             >
               Voir les conditions détaillées
-            </Text>
+            </Text> */}
             {order.status !== "delivered" && (
               <Flex
                 mt={14}