+++ webapp/src/pages/widget/index.tsx
@@ -223,14 +223,12 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
       },
     });
 
-    if (!context.req.cookies[process.env.NEXT_PUBLIC_WIDGET_TOKEN_NAME!]) {
-      context.res.setHeader(
-        "Set-Cookie",
-        `${process.env.NEXT_PUBLIC_WIDGET_TOKEN_NAME}=${widgetToken}; Expires=${new Date(
-          new Date().getTime() + 7 * 24 * 60 * 60 * 1000
-        ).toUTCString()}; Path=/; SameSite=None; Secure`
-      );
-    }
+    context.res.setHeader(
+      "Set-Cookie",
+      `${process.env.NEXT_PUBLIC_WIDGET_TOKEN_NAME}=${widgetToken}; Expires=${new Date(
+        new Date().getTime() + 7 * 24 * 60 * 60 * 1000
+      ).toUTCString()}; Path=/; SameSite=None; Secure`
+    );
 
     if (!!users.docs.length) {
       return {